Added pre5 sources to main branch..
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@1925 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ mplayer/osdep/timer.h
@@ -2,8 +2,8 @@
 #define __TIMER_H
 
 void InitTimer();
-unsigned long GetTimer();
-unsigned long GetTimerMS();
+unsigned int GetTimer();
+unsigned int GetTimerMS();
 //int uGetTimer();
 float GetRelativeTime();
